+++ lib/src/analyzers/lint_analyzer/rules/rules_list/no_equal_arguments/visitor.dart
@@ -39,14 +39,17 @@ class _Visitor extends RecursiveAstVisitor<void> {
             arg is NamedExpression &&
             argument.expression is! Literal &&
             arg.expression is! Literal) {
-          return argument.expression.toString() == arg.expression.toString();
+          return argument.expression.staticParameterElement ==
+                  arg.expression.staticParameterElement &&
+              argument.expression.toString() == arg.expression.toString();
         }
 
         if (_bothLiterals(argument, arg)) {
           return argument == arg;
         }
 
-        return argument.toString() == arg.toString();
+        return argument.staticParameterElement == arg.staticParameterElement &&
+            argument.toString() == arg.toString();
       });
 
       if (argument != lastAppearance) {