+++ test/src/analyzers/lint_analyzer/rules/rules_list/no_equal_arguments/no_equal_arguments_rule_test.dart
@@ -9,6 +9,8 @@ const _incorrectExamplePath =
     'no_equal_arguments/examples/incorrect_example.dart';
 const _namedParametersExamplePath =
     'no_equal_arguments/examples/named_parameters_example.dart';
+const _providerExamplePath =
+    'no_equal_arguments/examples/provider_example.dart';
 
 void main() {
   group('NoEqualArgumentsRule', () {
@@ -72,6 +74,13 @@ void main() {
       locationTexts: ['lastName: firstName'],
       messages: ['The argument has already been passed.'],
     );
+  });
+
+  test('reports no issues for provider read', () async {
+    final unit = await RuleTestHelper.resolveFromFile(_providerExamplePath);
+    final issues = NoEqualArgumentsRule().check(unit);
+
+    RuleTestHelper.verifyNoIssues(issues);
   });
 
   test('reports no issues with custom config', () async {